Updated skills.md and formatting issues
